+++ src/type/inspector/index.ts
