+++ src/type/type.ts
@@ -67,94 +67,9 @@ export interface DefinedTable extends IExtendedType {
     readonly fields: Map<string, TType>;
 }
 
-// export interface IExtendedType {
-//     readonly kind: TExtendedTypeKind;
-//     readonly maybeExtendedKind: ExtendedTypeKind;
-//     readonly isNullable: boolean;
-// }
-
-// export interface ExtendedRecord extends IExtendedType {
-//     readonly kind: TExtendedTypeKind;
-//     readonly maybeExtendedKind: ExtendedTypeKind;
-//     readonly isNullable: boolean;
-// }
-
-// function isExtendedType<T>(ttype: TType): ttype is TExtendedType {
-//     return ttype.maybeExtendedKind !== undefined;
-// }
-
-// function maybeAsExtendedType<T>(ttype: TType, extendedTypeKind: ExtendedTypeKind): undefined | (T & TExtendedType) {
-//     if (!isExtendedType(ttype)) {
-//         return undefined;
-//     }
-
-//     switch (extendedTypeKind) {
-//         case ExtendedTypeKind.Record:
-//             return;
-//     }
-// }
-
-// export interface ICustomType {
-//     readonly kind: TExtendedTypeKind;
-//     readonly isNullable: boolean;
-//     readonly isCustom: boolean;
-// }
-
-// export interface IRecord extends ICustomType {
-//     readonly kind: TypeKind.Record;
-//     readonly isCustom: boolean;
-// }
-
-// export interface RecordType extends IRecord {
-//     readonly isCustom: false;
-// }
-
-// export interface CustomRecordType extends IRecord {
-//     readonly isCustom: true;
-//     readonly fields: Map<string, undefined | TypeKind>;
-// }
-
-// export interface ITable extends ICustomType {
-//     readonly kind: TypeKind.Table;
-//     readonly isCustom: boolean;
-// }
-
-// export interface TableType extends ITable {
-//     readonly isCustom: false;
-// }
-
-// export interface CustomTableType extends ITable {
-//     readonly isCustom: true;
-//     readonly fields: Map<string, undefined | TypeKind>;
-// }
-
-// export interface IFunctionExpression extends IType {
-//     readonly kind: TypeKind.Function;
-//     readonly isEach: boolean;
-//     readonly isReturnNullable: boolean;
-//     readonly returnType: TypeKind;
-// }
-
-// export interface EachFunctionExpressionType extends IFunctionExpression {
-//     readonly isEach: true;
-//     readonly returnType: TypeKind;
-// }
-
-// export interface FunctionExpressionType extends IFunctionExpression {
-//     readonly isEach: false;
-//     readonly parameters: ReadonlyArray<FunctionParameter>;
-// }
-
-// export type TFunctionExpression = EachFunctionExpressionType | FunctionExpressionType;
-
-// // export type FunctionExpressionType = FunctionExpression & IType & { readonly kind: TypeKind.Function };
-
-// export interface FunctionParameter {
-//     readonly name: Ast.Identifier;
-//     readonly isOptional: boolean;
-//     readonly isNullable: boolean;
-//     readonly maybeType: Ast.TConstantKind | undefined;
-// }
+export interface ParameterType extends IExtendedType {
+    readonly isOptional: boolean;
+}
 
 export interface SimplifiedNullablePrimitiveType {
     readonly typeKind: TypeKind;
